// src/app/shared/models/goods.ts
export class Goods {
  offreur?: string;
  typeBien?: any;
  categoryBien?: any;
  atouts?: any;
  properties?: string;
  typeVisite?: any;
  titre?: string;
  description?: string;
  chambres?: number;
  capacite?: number;
  salleBains?: number;
  commodite?: string;
  regle?: string;
  serviceSuplementaire?: string;
  emplacement?: string;
  disponibilte?: String;
  photos?: any;
  pays?: any;
  zones?: any;
  tarifs?: string;
}